🚚 Move constant arbitraries into /arbitrary
