+++ src/check/arbitrary/LoremArbitrary.ts
@@ -1,5 +1,5 @@
 import { array } from '../../arbitrary/array';
-import { constant } from './ConstantArbitrary';
+import { constant } from '../../arbitrary/constant';
 import { Arbitrary } from './definition/Arbitrary';
 import { frequency } from '../../arbitrary/frequency';
