+++ src/check/arbitrary/ObjectArbitrary.ts
@@ -3,7 +3,7 @@ import { Arbitrary } from './definition/Arbitrary';
 import { stringify } from '../../utils/stringify';
 import { array } from '../../arbitrary/array';
 import { boolean } from './BooleanArbitrary';
-import { constant } from './ConstantArbitrary';
+import { constant } from '../../arbitrary/constant';
 import { dictionary, toObject } from './DictionaryArbitrary';
 import { double } from './FloatingPointArbitrary';
 import { frequency } from '../../arbitrary/frequency';
