+++ test/unit/check/arbitrary/FunctionArbitrary.spec.ts
@@ -1,7 +1,7 @@
 import * as fc from '../../../../lib/fast-check';
 
 import { func, compareFunc, compareBooleanFunc } from '../../../../src/check/arbitrary/FunctionArbitrary';
-import { context } from '../../../../src/check/arbitrary/ContextArbitrary';
+import { context } from '../../../../src/arbitrary/context';
 import { integer } from '../../../../src/arbitrary/integer';
 import { hasCloneMethod, cloneMethod } from '../../../../src/check/symbols';
 import { hash } from '../../../../src/utils/hash';
