+++ test/unit/check/arbitrary/ObjectArbitrary.spec.ts
@@ -1,7 +1,7 @@
 import * as prand from 'pure-rand';
 import * as fc from '../../../../lib/fast-check';
 
-import { constant } from '../../../../src/check/arbitrary/ConstantArbitrary';
+import { constant } from '../../../../src/arbitrary/constant';
 import { oneof } from '../../../../src/arbitrary/oneof';
 import {
   anything,
