+++ test/unit/check/arbitrary/RecordArbitrary.spec.ts
@@ -1,7 +1,7 @@
 import * as prand from 'pure-rand';
 import * as fc from '../../../../lib/fast-check';
 
-import { constant } from '../../../../src/check/arbitrary/ConstantArbitrary';
+import { constant } from '../../../../src/arbitrary/constant';
 import { Arbitrary } from '../../../../src/check/arbitrary/definition/Arbitrary';
 import { integer } from '../../../../src/arbitrary/integer';
 import { record, RecordConstraints } from '../../../../src/check/arbitrary/RecordArbitrary';
