+++ test/unit/check/arbitrary/StreamArbitrary.spec.ts
@@ -1,6 +1,6 @@
 import * as fc from '../../../../lib/fast-check';
 
-import { context } from '../../../../src/check/arbitrary/ContextArbitrary';
+import { context } from '../../../../src/arbitrary/context';
 import { nat } from '../../../../src/arbitrary/nat';
 import { infiniteStream } from '../../../../src/check/arbitrary/StreamArbitrary';
 import { Stream } from '../../../../src/stream/Stream';
