+++ test/unit/check/model/commands/CommandsArbitrary.spec.ts
@@ -3,7 +3,7 @@ import * as fc from '../../../../../lib/fast-check';
 
 import { Command } from '../../../../../src/check/model/command/Command';
 import { Random } from '../../../../../src/random/generator/Random';
-import { constant } from '../../../../../src/check/arbitrary/ConstantArbitrary';
+import { constant } from '../../../../../src/arbitrary/constant';
 import { commands } from '../../../../../src/check/model/commands/CommandsArbitrary';
 import { genericTuple } from '../../../../../src/arbitrary/genericTuple';
 import { nat } from '../../../../../src/arbitrary/nat';
